Relocate Environment interface
