+++ Slim/Http/Environment.php
@@ -9,7 +9,7 @@
 namespace Slim\Http;
 
 use Slim\Collection;
-use Slim\Interfaces\EnvironmentInterface;
+use Slim\Interfaces\Http\EnvironmentInterface;
 
 /**
  * Environment
