+++ Slim/Interfaces/Http/EnvironmentInterface.php
@@ -30,7 +30,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-namespace Slim\Interfaces;
+namespace Slim\Interfaces\Http;
 
 /**
  * Environment Interface
